+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2025-11-15
+
+### Added
+- **Get Content operation** in Message resource
+  - Download binary content (image, video, audio, file) from LINE messages
+  - Returns binary data that can be used with other nodes
+  - Use message ID from LINE Trigger events
+  - Supports custom binary property name
+
 ## [0.2.0] - 2025-11-15
 
 ### Added